refactor(made two packages, api and entity
