+++ src/main/java/com/greenfoxacademy/opal/kalendaryo/kalendaryo/model/api/AuthResponse.java
@@ -1,4 +1,4 @@
-package com.greenfoxacademy.opal.kalendaryo.kalendaryo.model;
+package com.greenfoxacademy.opal.kalendaryo.kalendaryo.model.api;
 
 public class AuthResponse {
     private long userId;
