+++ src/main/java/com/greenfoxacademy/opal/kalendaryo/kalendaryo/model/entity/AuthModel.java
@@ -1,4 +1,4 @@
-package com.greenfoxacademy.opal.kalendaryo.kalendaryo.model;
+package com.greenfoxacademy.opal.kalendaryo.kalendaryo.model.entity;
 
 import javax.persistence.*;
 import java.util.List;
