+++ src/main/java/com/greenfoxacademy/opal/kalendaryo/kalendaryo/model/entity/EventResponse.java
@@ -1,4 +1,4 @@
-package com.greenfoxacademy.opal.kalendaryo.kalendaryo.model;
+package com.greenfoxacademy.opal.kalendaryo.kalendaryo.model.entity;
 
 
 import com.fasterxml.jackson.annotation.JsonInclude;
