+++ src/main/java/com/greenfoxacademy/opal/kalendaryo/kalendaryo/model/entity/UserModel.java
@@ -1,6 +1,8 @@
-package com.greenfoxacademy.opal.kalendaryo.kalendaryo.model;
+package com.greenfoxacademy.opal.kalendaryo.kalendaryo.model.entity;
 
 import com.google.api.client.util.Base64;
+import com.greenfoxacademy.opal.kalendaryo.kalendaryo.model.entity.AuthModel;
+
 import java.security.SecureRandom;
 import javax.persistence.*;
 import java.util.List;
