+++ src/main/java/com/greenfoxacademy/opal/kalendaryo/kalendaryo/repository/AuthModelRepository.java
@@ -1,7 +1,6 @@
 package com.greenfoxacademy.opal.kalendaryo.kalendaryo.repository;
 
-import com.greenfoxacademy.opal.kalendaryo.kalendaryo.model.AuthModel;
-import com.greenfoxacademy.opal.kalendaryo.kalendaryo.model.UserModel;
+import com.greenfoxacademy.opal.kalendaryo.kalendaryo.model.entity.AuthModel;
 
 import java.util.List;
 import org.springframework.data.repository.CrudRepository;
