+++ src/main/java/com/greenfoxacademy/opal/kalendaryo/kalendaryo/repository/EventResponseRepository.java
@@ -1,6 +1,6 @@
 package com.greenfoxacademy.opal.kalendaryo.kalendaryo.repository;
 
-import com.greenfoxacademy.opal.kalendaryo.kalendaryo.model.EventResponse;
+import com.greenfoxacademy.opal.kalendaryo.kalendaryo.model.entity.EventResponse;
 import org.springframework.data.repository.CrudRepository;
 
 public interface EventResponseRepository extends CrudRepository<EventResponse, Long> {
