+++ src/main/java/com/greenfoxacademy/opal/kalendaryo/kalendaryo/repository/UserModelRepository.java
@@ -1,7 +1,7 @@
 package com.greenfoxacademy.opal.kalendaryo.kalendaryo.repository;
 
 
-import com.greenfoxacademy.opal.kalendaryo.kalendaryo.model.UserModel;
+import com.greenfoxacademy.opal.kalendaryo.kalendaryo.model.entity.UserModel;
 import org.springframework.data.repository.CrudRepository;
 
 public interface UserModelRepository extends CrudRepository<UserModel, Long> {
