+++ src/main/java/com/greenfoxacademy/opal/kalendaryo/kalendaryo/service/EventResponseService.java
@@ -1,6 +1,6 @@
 package com.greenfoxacademy.opal.kalendaryo.kalendaryo.service;
 
-import com.greenfoxacademy.opal.kalendaryo.kalendaryo.model.EventResponse;
+import com.greenfoxacademy.opal.kalendaryo.kalendaryo.model.entity.EventResponse;
 import com.greenfoxacademy.opal.kalendaryo.kalendaryo.repository.EventResponseRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
